+++ src/app/shared/interfaces/workflow/nestedsteps.ts
@@ -3,6 +3,7 @@ import { Options } from './options';
 export interface NestedSteps {
     label: string;
     name: string;
+    description: string;
     value: string;
     type: string;
     validators?: any;