+++ CHANGELOG.md
@@ -10,11 +10,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added 
 
-- 
+- Allows users to enter downstream distance in "Query by Fire Perimeters" workflow
+- Descriptions to workflows and steps
 
 ### Changed  
 
 - Split 'Select a Fire Perimeter' step in 'Query by Fire Perimeters' workflow into two seperate steps. The first is to select your fire perimeter the second it to start the trace. 
+- Updated layout
 
 ### Deprecated